+++ .github/workflows/test-actions.yml
@@ -64,11 +64,17 @@ jobs:
     - name: build
       env:
         ConnectionStrings__BooksDb: ${{ secrets.CONNECTION_STRING }}
+        Remote__Username: ${{ secrets.RemoteUsername }}
+        Remote__Password: ${{ secrets.RemotePassword }}
+        Remote__Server: "soc-app-liv-17.napier.ac.uk"
       run: dotnet build ./Pipeline/Pipeline/Pipeline.csproj --no-restore --framework net8.0
 
     - name: Test
       env:
         ConnectionStrings__BooksDb: ${{ secrets.CONNECTION_STRING }}
+        Remote__Username: ${{ secrets.RemoteUsername }}
+        Remote__Password: ${{ secrets.RemotePassword }}
+        Remote__Server: "soc-app-liv-17.napier.ac.uk"
       run: dotnet-coverage collect "dotnet test ./Pipeline/Pipeline.sln --framework net8.0" -f xml -o "coverage.xml"
 
     - name: End Sonar Analysis
